+++ agent/src/main/java/eu/cqse/teamscale/jacoco/agent/PreMain.java
@@ -1,31 +1,37 @@
 package eu.cqse.teamscale.jacoco.agent;
 
-import eu.cqse.teamscale.jacoco.util.LoggingUtils;
-import org.apache.logging.log4j.LogManager;
-
 import java.lang.instrument.Instrumentation;
+import java.lang.reflect.Method;
+import java.net.URL;
+import java.net.URLClassLoader;
 
 /** Container class for the premain entry point for the agent. */
 public class PreMain {
 
-	/** Entry point for the agent, called by the JVM. */
+	/**
+	 * Entry point for the agent, called by the JVM.
+	 * The only purpose of this method is to enforce classloader isolation of
+	 * the agent from the rest of the application by running the actual agent
+	 * in a custom classloader that only has access to the agent's Jar file.
+	 * <p>
+	 * We do this by temporarily changing the thread's context classloader to
+	 * an {@link URLClassLoader} without a parent, starting the agent and then
+	 * resetting the context classloader.
+	 */
 	public static void premain(String options, Instrumentation instrumentation) throws Exception {
-		AgentOptions agentOptions;
-		try {
-			agentOptions = AgentOptionsParser.parse(options);
-		} catch (AgentOptionParseException e) {
-			LoggingUtils.initializeDefaultLogging();
-			LogManager.getLogger(PreMain.class).fatal("Failed to parse agent options: " + e.getMessage(), e);
-			System.err.println("Failed to parse agent options: " + e.getMessage());
-			throw e;
-		}
+		ClassLoader contextClassLoader = Thread.currentThread().getContextClassLoader();
 
-		LoggingUtils.initializeLogging(agentOptions.getLoggingConfig());
+		URL agentJarUrl = PreMain.class.getProtectionDomain().getCodeSource().getLocation();
+		URLClassLoader classloader = new URLClassLoader(new URL[]{agentJarUrl}, null);
+		Thread.currentThread().setContextClassLoader(classloader);
 
-		LogManager.getLogger(Agent.class).info("Starting JaCoCo's agent");
-		org.jacoco.agent.rt.internal_c13123e.PreMain.premain(agentOptions.createJacocoAgentOptions(), instrumentation);
+		// we must use reflection to access the agent's methods because we don't share
+		// the agent's classloader
+		String agentName = PreMain.class.getPackage().getName() + ".AgentBase";
+		Class<?> agentClass = classloader.loadClass(agentName);
+		Method method = agentClass.getMethod("premain", String.class, Instrumentation.class);
+		method.invoke(null, options, instrumentation);
 
-		AgentBase agent = agentOptions.createAgent();
-		agent.registerShutdownHook();
+		Thread.currentThread().setContextClassLoader(contextClassLoader);
 	}
 }